[installer]: create a configmap with a simplified installation manifest
This can be used to uninstall the application
